Move VM Group controller to its own folder
