+++ controllers/controllers.go
@@ -15,6 +15,7 @@ import (
 	spq "github.com/vmware-tanzu/vm-operator/controllers/storagepolicyquota"
 	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachine"
 	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachineclass"
+	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachinegroup"
 	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachineimagecache"
 	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachinepublishrequest"
 	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachinereplicaset"
@@ -77,6 +78,12 @@ func AddToManager(ctx *pkgctx.ControllerManagerContext, mgr manager.Manager) err
 	// TODO: guard with a capability
 	if err := virtualmachinesnapshot.AddToManager(ctx, mgr); err != nil {
 		return fmt.Errorf("failed to initialize VirtualMachineSnapshot controller: %w", err)
+	}
+
+	if pkgcfg.FromContext(ctx).Features.VMGroups {
+		if err := virtualmachinegroup.AddToManager(ctx, mgr); err != nil {
+			return fmt.Errorf("failed to initialize VMG controller: %w", err)
+		}
 	}
 
 	return nil
