+++ controllers/virtualmachinegroup/virtualmachinegroup_controller_suite_test.go
@@ -12,7 +12,7 @@ import (
 
 	ctrlmgr "sigs.k8s.io/controller-runtime/pkg/manager"
 
-	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachine/virtualmachinegroup"
+	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachinegroup"
 	pkgcfg "github.com/vmware-tanzu/vm-operator/pkg/config"
 	pkgctx "github.com/vmware-tanzu/vm-operator/pkg/context"
 	"github.com/vmware-tanzu/vm-operator/test/builder"
