+++ docs/ref/api/v1alpha4.md
@@ -253,7 +253,7 @@ _Appears in:_
 GroupMember describes a member of a VirtualMachineGroup.
 
 _Appears in:_
-- [VirtualMachineGroupSpec](#virtualmachinegroupspec)
+- [VirtualMachineGroupBootOrderGroup](#virtualmachinegroupbootordergroup)
 
 | Field | Description |
 | --- | --- |
@@ -262,10 +262,6 @@ _Appears in:_
 VirtualMachine or VirtualMachineGroup.
 
 If omitted, it defaults to VirtualMachine. |
-| `powerOnDelay` _[Duration](https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.24/#duration-v1-meta)_ | PowerOnDelay is the amount of time to wait before powering on the member.
-
-If omitted, the member will be powered on immediately when the group's
-power state changes to PoweredOn. |
 
 ### GuestHeartbeatAction
 
@@ -1323,6 +1319,27 @@ _Appears in:_
 - [VirtualMachineCryptoStatus](#virtualmachinecryptostatus)
 
 
+### VirtualMachineGroupBootOrderGroup
+
+
+
+VirtualMachineGroupBootOrderGroup describes a boot order group within a
+VirtualMachineGroup.
+
+_Appears in:_
+- [VirtualMachineGroupSpec](#virtualmachinegroupspec)
+
+| Field | Description |
+| --- | --- |
+| `members` _[GroupMember](#groupmember) array_ | Members describes the names of VirtualMachine or VirtualMachineGroup
+objects that are members of this boot order group. The VM or VM Group
+objects must be in the same namespace as this group. |
+| `powerOnDelay` _[Duration](https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.24/#duration-v1-meta)_ | PowerOnDelay is the amount of time to wait before powering on all the
+members of this boot order group.
+
+If omitted, the members will be powered on immediately when the group's
+power state changes to PoweredOn. |
+
 ### VirtualMachineGroupMemberStatus
 
 
@@ -1388,9 +1405,13 @@ _Appears in:_
 | `groupName` _string_ | GroupName describes the name of the group that this group belongs to.
 
 If omitted, this group is not a member of any other group. |
-| `members` _[GroupMember](#groupmember) array_ | Members describes the names of VirtualMachine or VirtualMachineGroup
-objects that are members of this group. The VM or VM Group objects must
-be in the same namespace as this group. |
+| `bootOrders` _[VirtualMachineGroupBootOrderGroup](#virtualmachinegroupbootordergroup) array_ | BootOrders describes the boot sequence for this group members. Each boot
+order contains a set of members that will be powered on simultaneously,
+with an optional delay before powering on. The orders are processed
+sequentially in the order they appear in this list, with delays being
+cumulative across orders.
+
+When powering off, all members are stopped immediately without delays. |
 | `powerState` _[VirtualMachinePowerState](#virtualmachinepowerstate)_ | PowerState describes the desired power state of a VirtualMachineGroup.
 
 Please note this field may be omitted when creating a new VM group. This
@@ -1402,20 +1423,22 @@ However, once the field is set to a non-empty value, it may no longer be
 set to an empty value. This means that if the group's power state is
 PoweredOn, and a VM whose power state is PoweredOff is added to the
 group, that VM will be powered on. |
+| `nextForcePowerStateSyncTime` _string_ | NextForcePowerStateSyncTime may be used to force sync the power state of
+the group to all of its members, by setting the value of this field to
+"now" (case-insensitive).
+
+Please note it is not possible to schedule future syncs using this field.
+The only value that users may set is the string "now" (case-insensitive). |
 | `powerOffMode` _[VirtualMachinePowerOpMode](#virtualmachinepoweropmode)_ | PowerOffMode describes the desired behavior when powering off a VM Group.
 Refer to the VirtualMachine.PowerOffMode field for more details.
 
 Please note this field is only propagated to the group's members when
-the group's power state is changed.
-
-If omitted, the mode defaults to TrySoft. |
+the group's power state is changed. |
 | `suspendMode` _[VirtualMachinePowerOpMode](#virtualmachinepoweropmode)_ | SuspendMode describes the desired behavior when suspending a VM Group.
 Refer to the VirtualMachine.SuspendMode field for more details.
 
 Please note this field is only propagated to the group's members when
-the group's power state is changed.
-
-If omitted, the mode defaults to TrySoft. |
+the group's power state is changed. |
 
 ### VirtualMachineGroupStatus
 
@@ -3071,6 +3094,9 @@ full disks. The available modes are:
 - Online   -- Promote disks while the VM is powered on. VMs with
               snapshots do not support online promotion.
 - Offline  -- Promote disks while the VM is powered off.
+
+Please note, this field is ignored for encrypted VMs since they do not
+use delta disks.
 
 Defaults to Online. |
 | `bootOptions` _[VirtualMachineBootOptions](#virtualmachinebootoptions)_ | BootOptions describes the settings that control the boot behavior of the
